% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{boxplot_stats}
\alias{boxplot_stats}
\alias{boxplot_stats}
\title{Calculate boxplot stats}
\usage{
boxplot_stats(data, var)
}
\arguments{
\item{data}{tibble; Includes survival for each examined group and cancer.}

\item{var}{character; The variable to calculate the boxplot stats for. Must
be present in data.}
}
\value{
A tibble including the elements of the boxplot.
}
\description{
Calculates the boxplot stats to provide to ggplot as boxplot.stats and
stats_boxplot differ somewhat.
}
\author{
Jason T. Serviss
}
