% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QpcrFuns.R
\name{calcStats}
\alias{calcStats}
\alias{calcStats}
\title{calcStats}
\usage{
calcStats(data, cnt.group, cnt.value, grouping)
}
\arguments{
\item{data}{tibble; A tibble output by the \link{code{ddct}}
function.}
}
\value{
A tibble including statistics.
}
\description{
Calculates the mean, 95% confidence interval and t.test values.
}
\author{
Jason T. Serviss
}
