% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{coding_potential_cpat}
\alias{coding_potential_cpat}
\alias{coding_potential_cpat}
\title{Conding potential analysis via CPAT.}
\format{A tibble with 6 rows and 7 variables:
\describe{
  \item{sequenceName}{\url{http://rna-cpat.sourceforge.net}}
  \item{RNAsize}{\url{http://rna-cpat.sourceforge.net}}
  \item{ORFsize}{\url{http://rna-cpat.sourceforge.net}}
  \item{FicketScore}{\url{http://rna-cpat.sourceforge.net}}
  \item{hexamerScore}{\url{http://rna-cpat.sourceforge.net}}
  \item{codingProbability}{\url{http://rna-cpat.sourceforge.net}}
  \item{codingLabel}{\url{http://rna-cpat.sourceforge.net}}
  ...
}}
\source{
\url{http://lilab.research.bcm.edu/cpat/}

\url{http://cpc.cbi.pku.edu.cn}
}
\description{
Protein-coding capacity was evaluated using the Coding-potential Assessment
Tool \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3616698/} and
Coding-potential Calculator
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1933232/} with default
settings. Transcript sequences for use with Coding-potential Assessment Tool
were downloaded from the UCSC genome browser using the following IDs: HOTAIR
(ENST00000455246.1), XIST (ENST00000429829.1), β-actin (ENST00000331789.5),
Tubulin (ENST00000427480.1), and MYC (ENST00000377970). Transcript sequences
for use with Coding-potential Calculator were downloaded from the UCSC genome
browser using the following IDs: HOTAIR (uc031qho.1), β-actin (uc003soq.4).

Protein-coding capacity was evaluated using the Coding-potential Assessment
Tool \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3616698/} and
Coding-potential Calculator
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1933232/} with default
settings. Transcript sequences for use with Coding-potential Assessment Tool
were downloaded from the UCSC genome browser using the following IDs: HOTAIR
(ENST00000455246.1), XIST (ENST00000429829.1), β-actin (ENST00000331789.5),
Tubulin (ENST00000427480.1), and MYC (ENST00000377970). Transcript sequences
for use with Coding-potential Calculator were downloaded from the UCSC genome
browser using the following IDs: HOTAIR (uc031qho.1), β-actin (uc003soq.4).
}
\examples{
getData('Figure 1f')
getData('Figure 1-Supplement 2d')
}
