% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QpcrFuns.R
\name{dct}
\alias{dct}
\alias{dct}
\title{dct}
\usage{
dct(data, GOIs, HK, grouping, med = TRUE)
}
\arguments{
\item{data}{tibble; A tibble output by the \link{code{technicalMeans}}
function.}

\item{GOIs}{character. The genes of interest.}

\item{HK}{character. The house keeping gene.}

\item{grouping}{unquoted expression(s). Column names of variables that include the
experimental conditions.}

\item{logical.}{A logical indicating if the median of the house keeping genes
should be used to calculate dct; default = TRUE.}
}
\value{
A tibble including the dct values.
}
\description{
Calculates the delta ct value.
}
\author{
Jason T. Serviss
}
