% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QpcrFuns.R
\name{ddct}
\alias{ddct}
\alias{ddct}
\title{ddct}
\usage{
ddct(data, logical, grouping, med = TRUE)
}
\arguments{
\item{data}{tibble; A tibble output by the \link{code{dct}}
function.}

\item{logical}{tibble. A tibble with colnames equal to the colnames in the
data argument where the control values are located. The tibble should have
the same number of rows as the data argument and contain logical values
indicating if the corresponding value in data is a control value.}

\item{grouping}{character. Column names of variables that include the
experimental conditions.}

\item{logical.}{A logical indicating if the median of the control values
should be used to calculate ddct; default = TRUE.}
}
\value{
A tibble including the ddct values.
}
\description{
Calculates the delta delta ct value.
}
\author{
Jason T. Serviss
}
