% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\docType{methods}
\name{downloadTCGA}
\alias{downloadTCGA}
\alias{downloadTCGA}
\alias{downloadTCGA,logical-method}
\title{downloadTCGA}
\usage{
downloadTCGA(save, path, ...)

\S4method{downloadTCGA}{logical}(save = FALSE, path = NULL, ...)
}
\arguments{
\item{save}{logical. Should the data be saved locally.}

\item{path}{character. A path indicating where the data should be saved if
\code{save} is TRUE.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame.
}
\description{
downloadTCGA
}
\author{
Jason T. Serviss
}
\keyword{downloadTCGA}
