% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{figure1c}
\alias{figure1c}
\title{miR34a asRNA and miR34a TCGA correlation.}
\format{A tibble with 5823 rows and 7 variables:
\describe{
  \item{TCGA_cancer_id}{Character; TCGA sample ID.}
  \item{TP53}{Numeric; Indicates TP53 nonsynonymous mutation. 0 is wild type, 1 is mutated.}
  \item{RP3}{Numeric; Expression for miR34a asRNA.}
  \item{RP3_cna}{Numeric; Named numeric vector containing for each TCGA barcode information regarding RP3-510D11.2 copy number alteration (gene-based segmentation values). For details on processing, see *Ashouri A et al. Nat Commun.}
  \item{miR34a}{Numeric; Expression for miR34a.}
  \item{cancer}{Character; Indicates cancer type.}
  \item{PAM50}{Numeric; Named character vector containing for each TCGA barcode from the breast cancer dataset information regarding PAM50 classification: LumA, LumB, Basal or Her2. For details on PAM50 classification, see *Ashouri A et al. Nat Commun.}
  ...
}}
\description{
miR34a asRNA and miR34a TCGA correlation.
}
\examples{
getData('Figure 1c')
}
