% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{figure2a}
\alias{figure2a}
\title{miR34a asRNA expression in HCT116 and HEK293t upon doxorubicin treatment.}
\format{A tibble with 48 rows and 6 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{'Cell line'}{Character; Indicating the cell line the data corresponds to.}
  \item{Treatment}{Character; Indicates the treatment type.}
  \item{gene}{Character; indicates the gene that the Ct values correspond to.}
  \item{Ct1}{Numeric; Ct value corresponding to the first technical replicate.}
  \item{Ct2}{Numeric; Ct value corresponding to the 2nd technical replicate.}
  ...
}}
\description{
All cell lines were cultured at 5% CO2 and 37° C with HEK293T cells cultured
in DMEM high glucose (Hyclone) and HCT116 cells in McCoy’s 5a
(Life Technologies). All growth mediums were supplemented with 10%
heat-inactivated FBS and 50 μg/ml of streptomycin and 50 μg/ml of penicillin.
Cells were plated at 300,000 cells per well in a 6-well plate and cultured
overnight. The following day cells were treated with 0, 100, 200, or
500 ng/ml doxorubicin for 24hrs. RNA was extracted using the RNeasy mini kit
(Qiagen) and subsequently treated with DNase (Ambion Turbo DNA-free, Life
Technologies). 500ng RNA was used for cDNA synthesis using MuMLV (Life
Technologies) and a 1:1 mix of oligo(dT) and random nanomers. QPCR was
carried out using KAPA 2G SYBRGreen (Kapa Biosystems) using the Applied
Biosystems 7900HT machine with the cycling conditions: 95 °C for 3 min,
95 °C for 3 s, 60 °C for 30 s.
}
\examples{
getData('Figure 2a')
}
