% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{figure2b}
\alias{figure2b}
\title{miR34a asRNA doxorubicin treatment of HCT116 p53 wt and null.}
\format{A tibble with 72 rows and 7 variables:
\describe{
  \item{'Cell line'}{Character; Indicating the cell line the data corresponds to.}
  \item{Condition}{Factor; Indicating the genetic status of TP53 in the cell line.}
  \item{Treatment}{Factor; Indicates the treatment (doxorubicin) concentration (ng/ml).}
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{gene}{Factor; indicates the gene that the Ct values correspond to.}
  \item{Ct1}{Numeric; Ct value corresponding to the first technical replicate.}
  \item{Ct2}{Numeric; Ct value corresponding to the 2nd technical replicate.}
  ...
}}
\description{
HCT116 cells were cultured in DMEM high modified (Hyclone, GE healthcare)
supplemented with supplemented with 2mM L-glutamine, 50ug/ml
Penicillin-Streptomycin and 10% Fetal Calf Serum. 25 x 10^5 HCT116 cells
were plated in 6 well plates. 24 hours later media was exchanged and
doxorubicin was added to a final concentration of 100, 200 or 500 ng/ml.
Cells were harvested for RNA extraction 24 hours later using trypsin. RNA
was extracted using Nucleospin RNA kit (Machery-Nagel Ref. 740955) according
to manufacturer‟s protocol and DNase treated using Ambion Turbo DNA-free
according to manufacturer‟s protocol (Life Technologies Ref. AM1907). cDNA
was synthesized using ~500 ng RNA with M-MLV (Life Technologies Ref 28025013)
and a mixture of oligo(dT) with nanomers in accordance with the
manufacturer's protocol. qPCR quantification was carried out using the
PowerUp SYBR Green Master Mix (Thermo Fisher Scientific, Ref. A25777) on the
CFX96 Touch Real-Time PCR Detection System: 50°C for 2 min, 95°C for 2min,
and 95°C for 1 sec followed by 60°C for 30 sec repeated for 40 cycles.
}
\examples{
getData('Figure 2b')
}
