% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{figure2c}
\alias{figure2c}
\title{P1 transfections in HCT116 and HEK293t cells.}
\format{A tibble with 36 rows and 6 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{'Cell line'}{Character; Indicating the cell line the data corresponds to.}
  \item{gene}{Character; indicates the gene that the Ct values correspond to.}
  \item{alias}{Character; Indicating the genes alias.}
  \item{construct}{Character; Indicates the transfected construct.}
  \item{value}{Indicates the activity metric normalized to GFP.}
  ...
}}
\description{
All cell lines were cultured at 5% CO2 and 37° C with HEK293T cells cultured
in DMEM high glucose (Hyclone) and HCT116 cells in McCoy’s 5a
(Life Technologies). All growth mediums were supplemented with 10%
heat-inactivated FBS and 50 μg/ml of streptomycin and 50 μg/ml of penicillin.
Cells were plated at 10,000 cells per well in a 96-well plate with a white
bottom and cultured overnight. The following day cells were co-transfected
with 10ng of empty, p1, or p2 plasmid and GFP using the standard
lipofectamine 2000 (Life Technologies) protocol. The expression of GFP and
luminescence was measured 24 h post transfection by using the Dual-Glo
Luciferase Assay System (Promega) and detected by the GloMax-Multi+
Detection System (Promega). The expression of luminescence was normalized
to GFP.
}
\examples{
getData('Figure 2c')
}
