% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{figure2e}
\alias{figure2e}
\title{P1 transfections with shRNA Renilla and doxorubicin treatment.}
\format{A tibble with 72 rows and 6 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{treatment}{Factor; indicates the concentration of doxorubicin (ng/ml).}
  \item{shRNA}{Factor; Indicates the transfected shRNA.}
  \item{gene}{Factor; indicates the gene that the Ct values correspond to.}
  \item{Ct1}{Numeric; Ct value corresponding to the first technical replicate.}
  \item{Ct2}{Numeric; Ct value corresponding to the 2nd technical replicate.}
  ...
}}
\description{
We utilized the P1 construct where the overlapping region of miR34a HG and
miR34a AS is cloned with luciferase downstream of miR34a HG and renilla
downstream of miR34a AS. The p1 sequence was previously published in
Raver-Shapira, N., et al., Transcriptional activation of miR-34a contributes
to p53-mediated apoptosis. Mol Cell, 2007. 26(5): p. 731-43. All cell lines
were cultured at 5% CO2 and 37° C with HCT116 cells cultured in McCoy’s 5a
(Life Technologies). All growth mediums were supplemented with 10%
heat-inactivated FBS and 50 μg/ml of streptomycin and 50 μg/ml of penicillin.
2.5x10^5 HCT116wt cells were seeded in a 12-well plate. After 24hrs these
were co-transfected with the P1 construct (25ng) and shRenilla2.1 (250ng)
using lipo2000 (Life Technologies) standard protocol. 24hrs
post-transfection, doxorubicin treatment was initiated using 0, 300, or
500ng/ml. 24hrs post-treatment, RNA was extracted using the RNeasy mini kit
(Qiagen) and subsequently treated with DNase (Ambion Turbo DNA-free, Life
Technologies). 500ng RNA was used for cDNA synthesis using MuMLV (Life
Technologies) and a 1:1 mix of oligo(dT) and random nanomers. QPCR was
carried out using KAPA 2G SYBRGreen (Kapa Biosystems) using the Applied
Biosystems 7900HT machine with the cycling conditions: 95 °C for 3 min,
95 °C for 3 s, 60 °C for 30 s.
}
\examples{
getData('Figure 2e')
}
