% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{figure3c}
\alias{figure3c}
\title{Monitoring growth of miR34a asRNA overexpressings cells under conditions of
starvation.}
\format{A tibble with 1296 rows and 6 variables:
\describe{
  \item{'Cell line'}{Factor; Indicates the cell line the experiment was performed in.}
  \item{Treatment}{Factor; Indicates the cell line the experiment was performed in.}
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{'Technical Replicate'}{Factor; indicates the technical replicate.}
  \item{Time}{Character; Time in hours after treatment.}
  \item{Confluency}{Numeric; The percentage of confluency.}
  ...
}}
\description{
10^4 PC3 cells, either miR34a asRNA overexpressing or mock, were seeded in
96 well plates. After attachment (3-4h) media was replaced with either RPMI
(Gibco, life technology) (supplemented with 2mM L-glutamine, 50ug/ml
Penicillin-Streptomycin and 10% Fetal Calf Serum) or HBSS. Cells were
incubated in Spark Multimode Microplate reader for 48 hours at 37°C with 5%
CO2 in a humidity chamber. Confluency was measured every hour. Three
technical triplicates were included in each experiment and these were
independantly normalized for each condition to the 0 time point afterwhich
the mean was calculated. The 95% confidence interval was then calculated
based on the three independant experiments which were preformed.
}
\examples{
getData('Figure 3c')
}
