% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileMap}
\alias{fileMap}
\alias{fileMap}
\title{fileMap}
\usage{
fileMap(type)
}
\arguments{
\item{type, }{character; The type of source file to map. Can be Rmd, pdf, or
rds.}
}
\value{
list.
}
\description{
Maps figures to various source file types.
}
\examples{

fileMap()

}
\author{
Jason T. Serviss
}
