% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QpcrFuns.R
\name{folds}
\alias{folds}
\alias{folds}
\title{folds}
\usage{
folds(data)
}
\arguments{
\item{data}{tibble; A tibble output by the \link{code{ddct}}
function.}
}
\value{
A tibble including the fold and log2 fold values.
}
\description{
Calculates the fold and log2 fold values.
}
\author{
Jason T. Serviss
}
