% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getData}
\alias{getData}
\alias{getData}
\title{getData}
\usage{
getData(figure)
}
\arguments{
\item{figure}{Character; The name of the figure to get data for.}
}
\value{
tibble.
}
\description{
Gets package data.
}
\examples{

getData('Figure 2a')

}
\author{
Jason T. Serviss
}
