% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{hPlot}
\alias{hPlot}
\alias{hPlot}
\title{Plots the survival boxplot inset and p-value.}
\usage{
hPlot(survData, xVar, tRA = theme_remove_all)
}
\arguments{
\item{xVar}{character; The variable to be plotted on the x-axis.}

\item{tRA}{Extra ggplot2 theme elements.}

\item{data}{tibble; Includes survival for each examined group and cancer.}
}
\value{
A ggplot2 object.
}
\description{
Plots the survival boxplot inset and p-value.
}
\author{
Jason T. Serviss
}
