% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{linearStats}
\alias{linearStats}
\alias{linearStats}
\title{Calculate correlation stats}
\usage{
linearStats(data, xVar, yVar)
}
\arguments{
\item{data}{tibble; Includes survival for each examined group and cancer.}

\item{xVar}{character; The 1st group to calculate the correlation for. Must
be a column in data.}

\item{yVar}{character; The 2nd group to calculate the correlation for. Must
be a column in data.}
}
\value{
A tibble with the p.value and estimate for the correlation between
variables.
}
\description{
Calculates the p-value and r-value for the correlation shown in the upper
left hand side of the figure.
}
\author{
Jason T. Serviss
}
