% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lnc34a_cage}
\alias{lnc34a_cage}
\title{lnc34a CAGE analysis.}
\format{A tibble with 74 rows and 35 variables:}
\source{
\url{https://www.encodeproject.org}
}
\description{
All available CAGE data from 36 cell lines was downloaded from UCSC with the
script entitled "lnc34a CAGE" in the data-raw/saveRDS.R file. Of these 28
cell lines had CAGE transcription start sites mapping to the plus strand of
chromosome 1 and in regions corresponding to 200 bp upstream of the lnc34a
start site (9241796 - 200) and 200 bp upstream of the GENCODE annotated
miR34a asRNA start site (9242263 + 200). These cell lines included: AG04450,
BJ, GM12878, H1-hESC, HAoAF, HAoEC, HCH, HepG2, HFDPC, HMEpC, hMSC-AT,
hMSC-BM, hMSC-UC, HOB, HPC-PL, HPIEpC, HSaVEC, HUVEC, HVMF, HWP, IMR90,
MCF-7, NHDF, NHEK, NHEM.f_M2, NHEM_M2, SkMC, SK-N-SH_RA.
}
\examples{
getData('Supplementary Document 1a')
}
