% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{mainPlot}
\alias{mainPlot}
\alias{mainPlot}
\title{Plots the survival scatterplot.}
\usage{
mainPlot(survData, xVar, yVar, tRA = theme_remove_all)
}
\arguments{
\item{xVar}{character; The variable to be plotted on the x-axis.}

\item{yVar}{character; The variable to be plotted on the y-axis.}

\item{tRA}{Extra ggplot2 theme elements.}

\item{data}{tibble; Includes survival for each examined group and cancer.}
}
\value{
A ggplot2 object.
}
\description{
Plots the survival scatterplot.
}
\author{
Jason T. Serviss
}
