% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{p1_hek293t}
\alias{p1_hek293t}
\title{P1 shRenilla HEK293t transfections.}
\format{A tibble with 54 rows and 6 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{shRNA}{Factor; Indicates the transfected shRNA.}
  \item{belongsTo}{Character; indicates which shControls correspond to which shRNA.}
  \item{gene}{Factor; indicates the gene that the Ct values correspond to.}
  \item{Ct1}{Numeric; Ct value corresponding to the first technical replicate.}
  \item{Ct2}{Numeric; Ct value corresponding to the 2nd technical replicate.}
  ...
}}
\description{
We utilized the P1 construct where the overlapping region of miR34a HG and
miR34a AS is cloned with luciferase downstream of miR34a HG and renilla
downstream of miR34a AS. The p1 sequence was previously published in
Raver-Shapira, N., et al., Transcriptional activation of miR-34a contributes
to p53-mediated apoptosis. Mol Cell, 2007. 26(5): p. 731-43. All cell lines
were cultured at 5% CO2 and 37° C with HEK293T cells cultured in DMEM high
glucose (Hyclone). All growth mediums were supplemented with 10%
heat-inactivated FBS and 50 μg/ml of streptomycin and 50 μg/ml of penicillin.
2.5x10^5 293T cells were seeded in a 12-well plate. After 24hrs these were
co-transfected with 50 ng of the P1 construct and 250 ng shRenilla using
lipo2k standard protocol. After 48 hours, RNA was extracted using the RNeasy
mini kit (Qiagen) and subsequently treated with DNase (Ambion Turbo DNA-free,
Life Technologies). 500ng RNA was used for cDNA synthesis using MuMLV (Life
Technologies) and a 1:1 mix of oligo(dT) and random nanomers. QPCR was
carried out using KAPA 2G SYBRGreen (Kapa Biosystems) using the Applied
Biosystems 7900HT machine with the cycling conditions: 95 °C for 3 min,
95 °C for 3 s, 60 °C for 30 s.
}
\examples{
getData('Figure 2-Supplement 2')
}
