% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseUCSCfiles}
\alias{parseUCSCfiles}
\alias{parseUCSCfiles}
\title{parseUCSCfiles}
\usage{
parseUCSCfiles(url)
}
\arguments{
\item{url}{character. THe url of the metadata file.}
}
\value{
tibble.
}
\description{
Parse ENCODE metadata from UCSC. Currently only tested for CAGE.
}
\examples{

parseUCSCfiles()

}
\author{
Jason T. Serviss
}
