% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QpcrFuns.R
\name{pFormat}
\alias{pFormat}
\alias{pFormat}
\title{pFormat}
\usage{
pFormat(stats)
}
\arguments{
\item{data}{tibble; A tibble output by the \link{code{calcStats}}
function.}
}
\value{
A tibble including the formated p-values.
}
\description{
Formats the values for plotting and viewing.
}
\author{
Jason T. Serviss
}
