% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFigures.R
\name{plotFigure}
\alias{plotFigure}
\alias{plotFigure}
\title{plotFigure}
\usage{
plotFigure(figure, open = TRUE, outputPath = NULL, ...)
}
\arguments{
\item{figure}{Character; Figure to plot.}

\item{open}{Logical; Should the html file be opened in the local browser?}

\item{outputPath}{Character; Path to output the html file.}

\item{...}{additional arguments to pass on.}
}
\description{
Recreates any figure from scratch.
}
\details{
Include description.
}
\examples{

\dontrun{plotFigure("Figure 1a")}

}
\author{
Jason T. Serviss
}
