% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plotKM}
\alias{plotKM}
\alias{plotKM}
\title{Plots KM plots}
\usage{
plotKM(data, fit, gene, cancer, p.value)
}
\arguments{
\item{data}{tibble; A tibble including the class information.}

\item{fit}{survfit; A survfit object.}

\item{gene}{Unquoted variable indicating the gene being plotted.}

\item{cancer}{Unquoted variable indicating the cancer being plotted.}

\item{p.value}{numeric; The p-value corresponding to the gene and cancer
being plotted.}
}
\value{
Plots the results.
}
\description{
Designed to be run in pmap call.
}
\author{
Jason T. Serviss
}
