% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFigures.R
\name{renderFigure}
\alias{renderFigure}
\alias{renderFigure}
\title{renderFigure}
\usage{
renderFigure(figure)
}
\arguments{
\item{figure}{Character; Rmd to render}

\item{...}{additional arguments to pass on.}
}
\description{
Recreates any figure from scratch.
}
\details{
Include description.
}
\examples{

\dontrun{renderFigure("Figure 1a")}

}
\author{
Jason T. Serviss
}
