% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{renderPlots}
\alias{renderPlots}
\alias{renderPlots}
\title{Renders the scatter plot with the boxplot as an inset.}
\usage{
renderPlots(p1, p2)
}
\arguments{
\item{p1}{ggplot2; The scatter plot.}

\item{p2}{ggplot2; The box plot.}
}
\value{
A ggplot2 object.
}
\description{
Renders the scatter plot with the boxplot as an inset.
}
\author{
Jason T. Serviss
}
