% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stable_line_ccnd1_exp}
\alias{stable_line_ccnd1_exp}
\title{CCND1 expression in PC3 stable cell lines.}
\format{A tibble with 24 rows and 6 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{'Cell line'}{Character; indicates the cell line the experiment was performed in.}
  \item{condition}{Factor; Indicates the genetic modifications in the cell line.}
  \item{gene}{Character; indicates the gene that the Ct values correspond to.}
  \item{Ct1}{Numeric; Ct value corresponding to the first technical replicate.}
  \item{Ct2}{Numeric; Ct value corresponding to the 2nd technical replicate.}
  ...
}}
\description{
All cell lines were cultured at 5% CO2 and 37° C with PC3 cells in RPMI
(Hyclone) and 2 mM L-glutamine. All growth mediums were supplemented with
10% heat-inactivated FBS and 50 μg/ml of streptomycin and 50 μg/ml of
penicillin. Cells were seeded at 2x10^5 cells per well in a 12-well plate.
After 24 hours, RNA was extracted using the RNeasy mini kit (Qiagen) and
subsequently treated with DNase (Ambion Turbo DNA-free, Life Technologies).
500ng RNA was used for cDNA synthesis using MuMLV (Life Technologies) and a
1:1 mix of oligo(dT) and random nanomers. QPCR was carried out using KAPA 2G
SYBRGreen (Kapa Biosystems) using the Applied Biosystems 7900HT machine with
the cycling conditions: 95 °C for 3 min, 95 °C for 3 s, 60 °C for 30 s.
}
\examples{
getData('Figure 2d')
}
