% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stable_line_ccnd1_prot}
\alias{stable_line_ccnd1_prot}
\title{CCND1 protein expression in PC3 stable lines.}
\format{A tibble with 12 rows and 6 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{condition}{Factor; Indicates the genetic modifications in the cell line.}
  \item{protein}{Character; indicates the protein that the values correspond to.}
  \item{pixel}{Numeric; \url{https://imagej.nih.gov/ij/}.}
  \item{Intensity}{Numeric; \url{https://imagej.nih.gov/ij/}.}
  \item{background}{Numeric; \url{https://imagej.nih.gov/ij/}.}
  ...
}}
\description{
All cell lines were cultured at 5% CO2 and 37° C with PC3 cells in RPMI
(Hyclone) and 2 mM L-glutamine. All growth mediums were supplemented with
10% heat-inactivated FBS and 50 μg/ml of streptomycin and 50 μg/ml of
penicillin. We seeded three independent experiments of 150,000 cells/well in
a six well plate (PC3 mock vs. PC3 miR34a AS F4) and grew them for 24 hours.
Before harvesting cells were controlled for their confluency. Cells with a
confluence of 60-75%, were harvested using trypsin. Cell pellets were frozen
down at -80°C before lysed for Western Blot analysis. Cells were lysed in
RIPA Buffer and run on a 4-12% Bis-Tris-SDS Gel using MOPS running buffer.
Proteins were transferred onto a nitrocellulose membrane using iBlot Turbo
Blotting Device. Proteins were transferred for 7 minutes. Membranes were
blocked for 1 hour at room temperature in 5% milk. Membranes were incubated
with a cyclin D1 antibody (92G2 Rabbit mAb, Cell Signalling) overnight at
4°C. The membrane was incubated using a goat anti-rabbit antibody conjugated
to HRP for 1 hour at room temperature. Membranes were washed 3-times for
5min each in TBS-T. Membranes were developed using chemiluminescence. After
the picture was taken, the membrane was stripped using 0.4M NaOH solution for
20min. Blocking step was repeated and primary antibody against GAPDH was
incubated for overnight at 4°C. Consequently, the membrane was incubated
using a goat anti-rabbit antibody conjugated to HRP for 1 hour at room
temperature. Membranes were washed 3-times for 5min in TBS-T. Membranes were
developed using chemiluminescence.
}
\examples{
getData('Figure 3-Supplement 2b')
}
