% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stable_line_cell_cycle}
\alias{stable_line_cell_cycle}
\title{miR34a asRNA PC3 stable cell line cell cycle.}
\format{A tibble with 12 rows and 6 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{'Cell line'}{Character; Indicates the cell line the experiment was performed in.}
  \item{condition}{Factor; indicates the genetic modifications in the cell line.}
  \item{phase}{Factor; The cell cycle phase.}
  \item{value}{Numeric; The percentage of cells in the corresponding phase}.
  ...
}}
\description{
All cell lines were cultured at 5% CO2 and 37° C with Skov3 and PC3 cells in
RPMI (Hyclone) and 2 mM L-glutamine. All growth mediums were supplemented
with 10% heat-inactivated FBS and 50 μg/ml of streptomycin and 50 μg/ml of
penicillin. 1x10^5 cells per well in a 6-well plate and harvested after
24hrs. Cells were washed in PBS and fixed in 4% PFA at room temperature
overnight. PFA was removed, and cells were resuspended in 95% EtOH. The
samples were then rehydrated in distilled water, stained with DAPI and
analyzed by flow cytometry on a LSRII (BD Biosciences) machine. The cell
cycle phases ModFit software (Verity Software House) was used to quantify the
percentage of cells in each cell cycle phase. Percentages were converted to
log2 fractions for each sample. Student's t-test was used for statistical
analysis comparing the mock vs miR34a asRNA overespression conditions using
five (PC3) or three (Skov3) independant experiments.
}
\examples{
getData('Figure 3b')
}
