% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stable_line_expression}
\alias{stable_line_expression}
\title{miR34a asRNA stable over-expression cell lines.}
\format{A tibble with 120 rows and 7 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{'Cell line'}{Factor; Indicates the cell line the experiment was performed in.}
  \item{'Genetic mod'}{Factor; indicates the genetic modifications in the cell line.}
  \item{gene}{Character; indicates the gene that the Ct values correspond to.}
  \item{Ct1}{Numeric; The Ct value corresponding to the 1st technical triplicate.}
  \item{Ct2}{Numeric; The Ct value corresponding to the 2nd technical triplicate.}
  \item{Ct3}{Numeric; The Ct value corresponding to the 3rd technical triplicate.}
  ...
}}
\description{
RNA was extracted using the RNeasy mini kit (Qiagen) and subsequently
treated with DNase (Ambion Turbo DNA-free, Life Technologies). 500ng RNA was
used for cDNA synthesis using MuMLV (Life Technologies) and a 1:1 mix of
oligo(dT) and random nanomers. QPCR was carried out using KAPA 2G SYBRGreen
(Kapa Biosystems) using the Applied Biosystems 7900HT machine with the
cycling conditions: 95 °C for 3 min, 95 °C for 3 s, 60 °C for 30 s. QPCR for
miRNA expression analysis was performed according to the protocol for the
TaqMan microRNA Assay kit (Life Technologies) with primer/probe sets
purchased from the same company (TaqMan® MicroRNA Assay, hsa-miR-34a, human
and Control miRNA Assay, RNU48, human) and the same cycling scheme as above.
QPCR primers are shown below. Two experimental (technical) replicates were
included in each QPCR run and delta ct was calculated for each sample using
the mean of the gene of interest's technical replicates and the house keeping
gene's technical replicates. delta-delta ct was calculated for each sample
by subtracting the median dct value for the corresponding mock samples. Five
independant experiments were performed in total. The Student's t-test was
used to compare the mock vs miR34a asRNA overexpressing group's delta-delta
ct values for both genes.
}
\examples{
getData('Figure 3a')
}
