% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stable_line_pol2_chip}
\alias{stable_line_pol2_chip}
\title{PC3 stable line pollII ChIP.}
\format{A tibble with 16 rows and 7 variables:
\describe{
  \item{'Biological Replicate'}{Numeric; Indicates the biological replicate of the sample.}
  \item{'Cell line'}{Factor; Indicates the cell line the experiment was performed in.}
  \item{condition}{Factor; Indicates the genetic modification in the cell.}
  \item{sample}{Factor; indicates the sample type.}
  \item{gene}{Character; Gene amplified in QPCR.}
  \item{Quantity1}{Numeric; The absolute RNA value from technical replicate 1.}
  \item{Quantity2}{Numeric; The absolute RNA value from technical replicate 2.}
  ...
}}
\description{
QPCR was performed in technical replicate for each sample and the mean of the
technical replicates was subsequently used to calculate the fraction of
input. The fold change of the miR34a asRNA overexpression samples was then
calculated by dividing the values by the values for the corresponding mock
samples. The fold values were then log2 transformed and the Student's t-test
was used to test for significant differences under the null hypothesis that
the true mean value of log2 transformed fold changes were equal to 0.
}
\examples{
getData('Figure 3d')
}
