% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{supp_figure1a}
\alias{supp_figure1a}
\title{TCGA correlation table.}
\format{A tibble with 14 rows and 7 variables:
\describe{
  \item{Cancer}{Character; Indicates the tumor type.}
  \item{r_total}{Numeric; Spearman's rho for all samples in the cancer.}
  \item{p_value_total}{Numeric; P-value for all samples in the cancer.}
  \item{r_mutated}{Numeric; Sprearman's rho for TP53 mutated samples.}
  \item{p_value_mutated}{Numeric; P-value for mutated samples in the cancer.}
  \item{r_nonmut}{Numeric; Sprearman's rho for TP53 wild type samples.}
  \item{p_value_nonmut}{Numeric; P-value for TP53 wild type samples.}
  ...
}}
\description{
R-squared and p-values from the correlation analysis investigating the
correlation between miR34a and miR34a asRNA expression in p53 wild type (wt)
and mutated (mut) samples within TCGA cancer types.
}
\examples{
getData('Supplementary Figure 1a')
}
