% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{supp_figure2e}
\alias{supp_figure2e}
\title{Conding potential analysis via CPC.}
\format{A tibble with 3 rows and 10 variables:
\describe{
  \item{'Transcript name'}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{Coding}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{'Coding Potential Score'}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{'Hit number'}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{'Hit score'}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{'Frame score'}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{Length}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{'Coverage'}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{'Log-Odds score'}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  \item{Type}{\url{http://cpc.cbi.pku.edu.cn/docs/quick_guide.jsp}}
  ...
}}
\source{
\url{http://cpc.cbi.pku.edu.cn}
}
\description{
Protein-coding capacity was evaluated using the Coding-potential Assessment
Tool \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3616698/} and
Coding-potential Calculator
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1933232/} with default
settings. Transcript sequences for use with Coding-potential Assessment Tool
were downloaded from the UCSC genome browser using the following IDs: HOTAIR
(ENST00000455246.1), XIST (ENST00000429829.1), β-actin (ENST00000331789.5),
Tubulin (ENST00000427480.1), and MYC (ENST00000377970). Transcript sequences
for use with Coding-potential Calculator were downloaded from the UCSC genome
browser using the following IDs: HOTAIR (uc031qho.1), β-actin (uc003soq.4).
}
\examples{
getData('Supplementary Figure 2e')
}
