% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{supp_figure5c}
\alias{supp_figure5c}
\title{Monitoring proliferation of miR34a asRNA overexpressings cells under
conditions of starvation.}
\format{A tibble with 976785 rows and 9 variables:
\describe{
  \item{Type}{Character; Indicates if the sample is stained or an unstained control.}
  \item{Time}{Character; Indicates the time the sample was harvested.}
  \item{`Cell line`}{Character; indicates the cell line that the values correspond to.}
  \item{Condition}{Character; indicates the culturing conditions for the sample.}
  \item{`Technical replicate}{Numeric; Indicates the technical replicate.}
  \item{`Biological replicate`}{Numeric; Indicates the biological replicate.}
  \item{`FSC-A`}{Numeric; Value for forward scatter area.}
  \item{`BSC-A`}{Numeric; Value for back scatter area}
  \item{Violet}{Numeric; Fluorescence intensity for cellTrace Violet.}
  ...
}}
\description{
20^4 PC3 cells, either miR34a asRNA overexpressing or mock, were harvested and
stained in 1 ml PBS with 5uM CellTrace Violet (Invitrogen) for 5min and
subsequently seeded in 12 well plates. Time 0 measurements were taken once
cells had attatched and treatments with RPMI (Gibco, life technology) 10%
FBS, RPMI 0.1% FBS, or HBSS were simultaniously initiated in the remaining
cells. RPMI mediums were all additionally supplemented with 2mM L-glutamine
and 50ug/ml Penicillin-Streptomycin and 10% Fetal Calf Serum. Cells were
incubated for the indicated times before harvesting and CellTrace Violet was
quantified via flow cytometry (Sony SH800S Cell Sorter). Time 0 was performed
in biological triplicate and technical duplicate whereas all other time
points were performed in biological triplicates with one technical replicate.
}
\examples{
getData('Supplementary Figure 5c')
}
