% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{supp_figure7}
\alias{supp_figure7}
\title{lnc34a splice junction analysis.}
\format{A tibble with 373 rows and 10 variables:}
\source{
\url{https://www.encodeproject.org}
}
\description{
All available whole cell RNAseq splice junction data from the ENCODE project
originating from the Cold Spring Harbor Lab from 38 cell lines was downloaded
from the UCSC genome browser with the script entitled "lnc34a splice
junctions" in the data-raw/saveRDS.R file. Of these cell lines, 36 had
spliced reads mapping to the plus strand of chromosome 1 and in the region
between the lnc34a start (9241796) and transcription termination (9257102)
site (note that miR34a asRNA resides totally within this region). Splice
junctions from the following cell lines were included in the final figure:
A549, Ag04450, Bj, Cd20, Cd34mobilized, Gm12878, H1hesc, Haoaf, Haoec, Hch,
Helas3, Hepg2, Hfdpc, Hmec, Hmepc, Hmscat, Hmscbm, Hmscuc, Hob, Hpcpl,
Hpiepc, Hsavec, Hsmm, Huvec, Hvmf, Hwp, Imr90, Mcf7, Monocd14, Nhdf, Nhek,
Nhemfm2, Nhemm2, Nhlf, Skmc, Sknsh. All splice junctions were plotted and
colored according to the number of reads corresponding to each.
}
\examples{
getData('Supplementary Figure 7')
}
