% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_fit}
\alias{survival_fit}
\title{Calculate survival fit}
\arguments{
\item{surv_obj}{Surv; Object of class Surv.}

\item{class}{character; The classes if the survival should be a function of
class. Otherwise a function of 1.}
}
\value{
The survival fit.
}
\description{
Calculate survival fit
}
\author{
Jason T. Serviss
}
