% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_p}
\alias{survival_p}
\alias{survival_p}
\title{Calculate survival p-value}
\usage{
survival_p(surv_obj, class = NULL)
}
\arguments{
\item{surv_obj}{Surv; Object of class Surv.}

\item{class}{character; The classes in the Surv object.}
}
\value{
The survival fit.
}
\description{
Calculate survival p-value
}
\author{
Jason T. Serviss
}
