% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QpcrFuns.R
\name{technicalMeans}
\alias{technicalMeans}
\alias{technicalMeans}
\title{technicalMeans}
\usage{
technicalMeans(data, grouping)
}
\arguments{
\item{data}{tibble; See function description.}

\item{grouping}{character. See function description.}
}
\value{
A tibble including the values of the mean ct values.
}
\description{
Calculates the means of the two technical replicates.
}
\details{
Accepts a tibble with grouping variables and ct values (typically with one
column per ct value replicate). The grouping variable should include the
names of all columns where the ct values are not located.
}
\author{
Jason T. Serviss
}
