% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_survival}
\alias{get_survival}
\alias{get_survival}
\title{Calculate survival probability}
\usage{
get_survival(surv_data, n_years = 5)
}
\arguments{
\item{surv_data}{tibble; Includes columns vitalStatus and FU, date of last
follow-up.}

\item{n_years;}{numeric Years to examine survival.}
}
\value{
The survival probability.
}
\description{
Calculate survival probability
}
\author{
Jason T. Serviss
}
